fix(desktop): unblank renderer — break API client import cycle + relax dev CSP
The dev window rendered black, then (once scripts ran) crashed with
"Cannot access 'createClientConfig' before initialization" at client.gen.ts.

Two independent causes:

1. Import cycle / TDZ. hey-api's generated client.gen.ts imports
   `createClientConfig` from the runtimeConfigPath file, and that file imported
   `client` back. With the renderer importing runtime.ts first, runtime.ts's
   top-level `import { client }` forced client.gen.ts to evaluate before
   `createClientConfig` was initialized. Split the seam: a new cycle-free
   client-config.ts holds createClientConfig + config state (imports only
   token-store, never `client`); runtime.ts keeps configureClient and the live
   `client` wiring. Repoint runtimeConfigPath + the generated import at it.
   Consumers still import configureClient from '@nimi/contract/api/runtime'.

2. Dev CSP. script-src 'self' refused @vitejs/plugin-react's inline Refresh
   preamble + Vite's eval-based HMR, so React never mounted. Relax DEV_CSP only
   (unsafe-inline/unsafe-eval + ws/http localhost). The shipped app keeps the
   strict production CSP.
