+++ packages/contract/openapi-ts.config.ts
@@ -9,8 +9,10 @@ import { defineConfig } from '@hey-api/openapi-ts'
  *
  * The fetch client is bundled into @hey-api/openapi-ts (v0.73+), so no separate
  * client package is installed. Runtime config (base URL, auth) lives in
- * `src/api/runtime.ts` via `runtimeConfigPath` — that file survives
- * regeneration, the generated code does not.
+ * `src/api/client-config.ts` via `runtimeConfigPath` — that file survives
+ * regeneration, the generated code does not. It must NOT import the generated
+ * `client` (that would create an init cycle); the live-client wiring lives in
+ * `src/api/runtime.ts`.
  *
  * Paths are relative to this package (`packages/contract`).
  */
@@ -23,7 +25,7 @@ export default defineConfig({
   plugins: [
     {
       name: '@hey-api/client-fetch',
-      runtimeConfigPath: './src/api/runtime.ts'
+      runtimeConfigPath: './src/api/client-config.ts'
     }
   ]
 })
